requirements.txt
Flask==3.1.2
flask_cors==6.0.2
gensim==4.4.0
httpx==0.28.1
langchain_core==1.2.7
langchain_openai==1.1.7
langgraph==1.0.6
lightgbm==4.6.0
matplotlib==3.10.8
numpy==2.2.6
openai==2.15.0
pandas==2.3.3
python-dotenv==1.2.1
pytz==2025.2
Requests==2.32.5
scikit_learn==1.7.1
scipy==1.15.3
ddgs==9.10.0
azure-storage-blob==12.28.0
azure-identity==1.25.1

yfinance
line-bot-sdk